+++ CHANGELOG.md
@@ -13,6 +13,7 @@
 * Add `REFUND_FAILED` to WebhookNotification Kind
 * Add `finalCapture` to `TransactionRequest`
 * Deprecate `getPayPalTrackingId` in favor of `getPayPalTrackerId` in PackageDetails
+* Add `shippingTaxAmount` to `Transaction` and `TransactionRequest`
 
 ## 3.34.0
 * Add domains parameter support to `ClientTokenRequest`